requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.3
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.1.0
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
confection==0.0.4
configobj==5.0.8
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
docker==6.0.1
dpath==2.1.5
dulwich==0.21.3
dvc==2.54.0
dvc-data==0.47.2
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.3.1
dvc-studio-client==0.7.0
dvc-task==0.2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.2.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
funcy==2.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
greenlet==2.0.2
gunicorn==20.1.0
htmlmin==0.1.12
hydra-core==1.3.2
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mlflow==2.2.2
multidict==6.0.4
multimethod==1.9.1
murmurhash==1.0.9
mypy-extensions==1.0.0
nanotime==0.5.2
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.8.10
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
pygit2==1.12.0
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
querystring-parser==1.2.4
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.0
scipy==1.9.3
scmrepo==1.0.2
seaborn==0.12.2
Send2Trash==1.8.0
shap==0.41.0
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.9
sqlparse==0.4.4
sqltrie==0.3.1
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typer==0.7.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
vine==5.0.0
visions==0.7.5
voluptuous==0.13.1
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
xgboost==1.3.3
yarl==1.8.2
ydata-profiling==4.1.2
zc.lockfile==3.0.post1
zipp==3.15.0
